+++ core/templatetags/admin_data_tags.py
@@ -6,8 +6,6 @@ from django.template.defaulttags import register
 from django.urls.base import reverse
 from django.utils.safestring import mark_safe
 
-from data.models.suggestion import SuggestionGroupe
-
 logger = logging.getLogger(__name__)
 
 
@@ -184,16 +182,6 @@ def quote_filter(value):
 
 
 @register.filter
-def get_item(dictionary, key):
-    return dictionary.get(key)
-
-
-@register.filter
-def is_not_editable(key):
-    return key in SuggestionGroupe.NOT_EDITABLE_FIELDS
-
-
-@register.filter
 def display_diff_values(old_value, new_value):
     if not new_value:
         return old_value
@@ -202,8 +190,8 @@ def display_diff_values(old_value, new_value):
     return diff_display(old_value, new_value)
 
 
-@register.inclusion_tag("data/_partials/helper_updated_values.html")
-def helper_updated_values(field, value):
+@register.inclusion_tag("data/_partials/extra_links.html")
+def extra_links(field, value):
     if field in ["siren", "siret"]:
         url_map = {
             "siren": f"https://annuaire-entreprises.data.gouv.fr/entreprise/{value}",
