+++ static/to_compile/js/solution_map.ts
@@ -59,7 +59,6 @@ export class SolutionMap {
       sources.osm = {
         type: "raster",
         tiles: [
-          "https://tile.openstreetmap.org/{z}/{x}/{y}.png",
           "https://a.tile.openstreetmap.org/{z}/{x}/{y}.png",
           "https://b.tile.openstreetmap.org/{z}/{x}/{y}.png",
           "https://c.tile.openstreetmap.org/{z}/{x}/{y}.png",
